requirements.txt
alembic==1.16.2
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
migrate==0.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
six==1.17.0
spotipy==2.25.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
